requirements.txt
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
sentencepiece>=0.1.99
datasets>=2.12.0
accelerate>=0.20.0
safetensors>=0.3.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
